fix: correct default value backfill during AddField
