+++ internal/proxy/impl.go
@@ -842,7 +842,6 @@ func (node *Proxy) BatchDescribeCollection(ctx context.Context, request *milvusp
 				CollectionName: collectionName,
 			}
 		}
-
 		responses = append(responses, describeCollectionResponse)
 	}
