+++ internal/util/importutilv2/csv/row_parser.go
@@ -28,9 +28,9 @@ import (
 	"github.com/milvus-io/milvus/internal/util/importutilv2/common"
 	"github.com/milvus-io/milvus/internal/util/nullutil"
 	pkgcommon "github.com/milvus-io/milvus/pkg/v2/common"
-	"github.com/milvus-io/milvus/pkg/v2/util/funcutil"
 	"github.com/milvus-io/milvus/pkg/v2/util/merr"
 	"github.com/milvus-io/milvus/pkg/v2/util/parameterutil"
+	"github.com/milvus-io/milvus/pkg/v2/util/timestamptz"
 	"github.com/milvus-io/milvus/pkg/v2/util/typeutil"
 )
 
@@ -446,7 +446,7 @@ func (r *rowParser) parseEntity(field *schemapb.FieldSchema, obj string, useElem
 		}
 		return wkbValue, nil
 	case schemapb.DataType_Timestamptz:
-		tz, err := funcutil.ValidateAndReturnUnixMicroTz(obj, r.timezone)
+		tz, err := timestamptz.ValidateAndReturnUnixMicroTz(obj, r.timezone)
 		if err != nil {
 			return nil, err
 		}
